requirements.txt
tiktoken 
openai 
python-dotenv 
beautifulsoup4 
lxml
google-generativeai
streamlit